+++ public/packages/types/types/config/app.d.ts
@@ -1,0 +1,4 @@
+export interface App {
+  localURL?: string,
+  ports?: string[]
+}