+++ public/packages/types/types/config/image.d.ts
@@ -1,0 +1,9 @@
+// Names of supported tools https://www.bit-ship.dev/tools
+import {ToolsNames} from "../shared";
+
+export interface Image {
+  name: string,
+  dependencies?: {
+    [key in ToolsNames]?: string
+  }
+}