+++ public/packages/types/types/config/job.d.ts
@@ -1,0 +1,8 @@
+export interface Job {
+  on: {
+    commit?: {
+      on:  'pre-commit' | 'post-commit'
+    },
+  },
+  tasks: Array<string | string[]>
+}