+++ public/packages/types/types/config/task.d.ts
@@ -1,0 +1,8 @@
+export interface Task {
+  script: string,
+  location?: string,
+  env? : {
+    [key: string]: string
+  }
+  volumes?: string[]
+}